fixed crc32c on 64bit platforms (fixes crc errors)
git-svn-id: svn+ssh://svn.lighttpd.net/lighttpd/branches/lighttpd-merge-1.4.x@745 152afb58-edef-0310-8abb-c4023f1b3aa9

+++ src/crc32.h
@@ -2,7 +2,8 @@
 #define __crc32cr_table_h__
 
 #include <sys/types.h>
+#include <stdint.h>
 
-unsigned long generate_crc32c(char *string, size_t length);
+uint32_t generate_crc32c(char *string, size_t length);
 
 #endif
